requirements.txt
agate==1.9.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asn1crypto==1.5.1
attrs==26.1.0
babel==2.18.0
boto3==1.43.1
botocore==1.43.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
daff==1.4.2
dbt-adapters==1.22.10
dbt-common==1.37.5
dbt-core==1.11.8
dbt-extractor==0.6.0
dbt-protos==1.0.474
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.11.4
deepdiff==8.6.2
fastapi==0.136.1
filelock==3.29.0
h11==0.16.0
idna==3.13
importlib_metadata==8.9.0
iniconfig==2.3.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.1
keyring==25.7.0
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
orderly-set==5.5.0
packaging==26.2
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.9.6
pluggy==1.6.0
protobuf==6.33.6
py4j==0.10.9.7
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.1.0
pyspark==3.5.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
s3transfer==0.17.0
six==1.17.0
snowflake-connector-python==4.4.0
snowflake-sqlalchemy==1.9.0
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
text-unidecode==1.3
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
zipp==3.23.1
